+++ src/main/java/com/mava/ordemCulto/domain/oportunidades/dto/OportunidadeDTO.java
@@ -3,10 +3,10 @@ package com.mava.ordemCulto.domain.oportunidades.dto;
 import com.mava.ordemCulto.domain.oportunidades.MomentoOportunidade;
 
 public record OportunidadeDTO(
-        Integer id,
+        Long id,
         String nomePessoa,
         MomentoOportunidade momentoOportunidade,
-        Integer cultoId
+        Long cultoId
 ) {
 
 }
