+++ src/main/java/com/mava/ordemCulto/repositories/AvisosRepository.java
@@ -5,15 +5,14 @@ import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
-public interface AvisosRepository extends JpaRepository<Avisos, Integer> {
+public interface AvisosRepository extends JpaRepository<Avisos, Long> {
     @Query(value = "SELECT COUNT(*) > 0 FROM avisos WHERE culto_id = :cultoId", nativeQuery = true)
-    public boolean buscarRegistroExistente(Integer cultoId);
+    public boolean buscarRegistroExistente(Long cultoId);
 
     @Modifying
     @Transactional
     @Query("DELETE FROM Avisos a WHERE a.culto.id = :cultoId")
-    void deleteByCultoId(Integer cultoId);
+    void deleteByCultoId(Long cultoId);
 
 }
