+++ src/main/java/com/mava/ordemCulto/repositories/EquipeIntercessaoRepository.java
@@ -5,14 +5,13 @@ import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
-public interface EquipeIntercessaoRepository extends JpaRepository<EquipeIntercessao, Integer> {
+public interface EquipeIntercessaoRepository extends JpaRepository<EquipeIntercessao, Long> {
     @Query(value = "SELECT COUNT(*) > 0 FROM avisos WHERE culto_id = :cultoId", nativeQuery = true)
-    public boolean buscarRegistroExistente(Integer cultoId);
+    public boolean buscarRegistroExistente(Long cultoId);
 
     @Modifying
     @Transactional
     @Query("DELETE FROM EquipeIntercessao e WHERE e.culto.id = :cultoId")
-    void deleteByCultoId(Integer cultoId);
+    void deleteByCultoId(Long cultoId);
 }
