+++ src/main/java/com/mava/ordemCulto/repositories/OportunidadesRepository.java
@@ -4,8 +4,8 @@ import com.mava.ordemCulto.domain.oportunidades.Oportunidades;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface OportunidadesRepository extends JpaRepository<Oportunidades, Integer> {
+public interface OportunidadesRepository extends JpaRepository<Oportunidades, Long> {
     @Query(value = "SELECT COUNT(*) > 0 FROM avisos WHERE culto_id = :cultoId", nativeQuery = true)
-    public boolean buscarRegistroExistente(Integer cultoId);
+    public boolean buscarRegistroExistente(Long cultoId);
 
 }
